refactor(filesystem): reorganize flat module into domain-grouped subdirectories
Split the flat src/ layout into content/, formats/, and tree/ subdirectories.
Extract errors, ids, table, and path into their own top-level modules.
Barrel index re-exports preserve the existing public API.
