+++ packages/filesystem/src/content/index.ts
@@ -1,0 +1,13 @@
+export {
+	type ContentHelpers,
+	createContentHelpers,
+} from './content.js';
+export type {
+	BinaryEntry,
+	ContentMode,
+	RichTextEntry,
+	SheetEntry,
+	TextEntry,
+	TimelineEntry,
+} from './entry-types.js';
+export { createTimeline, type Timeline } from './timeline.js';
