--- packages/filesystem/src/file-table.ts
@@ -1,21 +1,0 @@
-import { defineTable } from '@epicenter/workspace';
-import { type } from 'arktype';
-import { FileId } from './types.js';
-
-export const filesTable = defineTable(
-	type({
-		id: FileId,
-		name: 'string',
-		parentId: FileId.or(type.null),
-		type: "'file' | 'folder'",
-		size: 'number',
-		createdAt: 'number',
-		updatedAt: 'number',
-		trashedAt: 'number | null',
-		_v: '1',
-	}),
-).withDocument('content', {
-	guid: 'id',
-	updatedAt: 'updatedAt',
-	tags: ['persistent'],
-});
