+++ packages/filesystem/src/tree/index.ts
@@ -1,0 +1,10 @@
+export {
+	assertUniqueName,
+	disambiguateNames,
+	validateName,
+} from './naming.js';
+export {
+	createFileSystemIndex,
+	type FileSystemIndex,
+} from './path-index.js';
+export { FileTree } from './tree.js';
