+++ packages/filesystem/src/tree/naming.ts
@@ -1,51 +1,7 @@
 import type { TableHelper } from '@epicenter/workspace';
-import type { FileId, FileRow } from './types.js';
-
-export type FsErrorCode =
-	| 'ENOENT'
-	| 'EISDIR'
-	| 'EEXIST'
-	| 'ENOSYS'
-	| 'EINVAL'
-	| 'ENOTEMPTY'
-	| 'ENOTDIR';
-
-/** Create an errno-style error with code property */
-function fsError(
-	code: FsErrorCode,
-	message: string,
-): Error & { code: FsErrorCode } {
-	const err = new Error(`${code}: ${message}`) as Error & { code: FsErrorCode };
-	err.code = code;
-	return err;
-}
-
-/**
- * Namespace of errno-style error factories for the virtual filesystem.
- *
- * Each method creates an `Error` with a `.code` property and a message
- * formatted as `"CODE: message"`. Type `FS_ERRORS.` to browse all
- * available codes via dot-autocomplete.
- *
- * @example
- * ```typescript
- * throw FS_ERRORS.ENOENT('/missing.txt');
- * throw FS_ERRORS.EISDIR('/some/dir');
- * throw FS_ERRORS.ENOSYS('symlinks not supported');
- * ```
- */
-export const FS_ERRORS = {
-	ENOENT: (message: string) => fsError('ENOENT', message),
-	EISDIR: (message: string) => fsError('EISDIR', message),
-	EEXIST: (message: string) => fsError('EEXIST', message),
-	ENOSYS: (message: string) => fsError('ENOSYS', message),
-	EINVAL: (message: string) => fsError('EINVAL', message),
-	ENOTEMPTY: (message: string) => fsError('ENOTEMPTY', message),
-	ENOTDIR: (message: string) => fsError('ENOTDIR', message),
-} satisfies Record<
-	FsErrorCode,
-	(message: string) => Error & { code: FsErrorCode }
->;
+import { FS_ERRORS } from '../errors.js';
+import type { FileId } from '../ids.js';
+import type { FileRow } from '../table.js';
 
 /** Validate a filename. Rejects path separators, null bytes, and reserved names. */
 export function validateName(name: string): void {
@@ -94,12 +50,12 @@ export function disambiguateNames(rows: FileRow[]): Map<string, string> {
 
 	for (const [name, group] of byName) {
 		if (group.length === 1) {
-			result.set(group[0]?.id, name);
+			result.set(group[0]!.id, name);
 			continue;
 		}
 		// Sort by createdAt — earliest keeps clean name
 		group.sort((a, b) => a.createdAt - b.createdAt);
-		result.set(group[0]?.id, name);
+		result.set(group[0]!.id, name);
 		for (let i = 1; i < group.length; i++) {
 			const row = group[i]!;
 			const dotIndex = name.lastIndexOf('.');
